+++ example/lib/src/full_player_example_page.dart
@@ -20,7 +20,9 @@ class FullPlayerExamplePage extends StatefulWidget {
 
 class _FullPlayerExamplePageState extends State<FullPlayerExamplePage> {
   late final VlcPlayerController _controller = VlcPlayerController(
-    source: Uri.parse('https://media.w3.org/2010/05/sintel/trailer.mp4'),
+    mediaSource: VlcMediaSource(
+      uri: Uri.parse('https://media.w3.org/2010/05/sintel/trailer.mp4'),
+    ),
     options: widget.playerOptions,
   );